requirements.txt
Flask==2.3.3
Pillow==10.0.1
rembg==2.0.50
numpy==1.24.3
opencv-python-headless==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
cloudinary==1.36.0
super-image==0.1.7
Werkzeug==2.3.7